fix(ci): remove unnecessary u64 casts in tui/app.rs
Clippy 1.95 flags `as u64` on fields that are already u64.

Made-with: Cursor

+++ rust/src/tui/app.rs
@@ -49,10 +49,10 @@ impl AppState {
             events: Vec::new(),
             total_saved: store
                 .total_input_tokens
-                .saturating_sub(store.total_output_tokens) as u64,
-            total_original: store.total_input_tokens as u64,
+                .saturating_sub(store.total_output_tokens),
+            total_original: store.total_input_tokens,
             cache_hits: store.cep.total_cache_hits,
-            total_calls: store.total_commands as u64,
+            total_calls: store.total_commands,
             files: std::collections::HashMap::new(),
             gain_score: None,
             last_gain_refresh: Instant::now(),
